+++ docs/RELEASE-NOTES.md
@@ -6,6 +6,7 @@ These are the release notes for the [Maps extension](../README.md).
 * Removed support for the deprecated Google JavaScript API
 * Updated the translations to use the new MediaWiki JSON format
 * Re added support for fill color and fill opacity parameters for circles
+* Re added image overlay support for Google Maps
 
 ## Maps 3.0.1